requirement.txt
Django==1.6.2
South==1.0.2
django-braces==1.8.1
django-crispy-forms==1.5.2
six==1.10.0
